move scripts out of v2 folder
