+++ scripts/manifest_logic.md
