+++ tests/v2/test_wiki.py
@@ -1,7 +1,7 @@
-from scripts.v2.wiki import *      # functions that we run tests on
-from scripts.v2.docker_adapter import *
-from scripts.v2.tree import *
-from scripts.v2.fs import *
+from scripts.wiki import *      # functions that we run tests on
+from scripts.docker_adapter import *
+from scripts.tree import *
+from scripts.fs import *
 
 from unittest.mock import MagicMock, patch
 import unittest
